Summary
This is a static web template made with HTML, CSS, and Bootstrap.
Because it contains only necessary files, it is a lightweight and fast-working website. It is also fully responsive.
Combined with simplicity, the theme appears serious, authoritative, and yet intuitive. It expertly balances text and visuals, giving each their due. The tagline is so large that it draws attention right away.